+++ lowcar/devices/KoalaBear/PID.cpp
@@ -11,19 +11,19 @@ PID::PID() {
     this->velocity = 0.0;
 
     // create and populate the prev_error and prev_time arrays
-    this->prev_error = new float[SAMPLES];
-    this->prev_time = new float[SAMPLES];
+    this->prev_error = new double[SAMPLES];
+    this->prev_time = new double[SAMPLES];
     for (int i = 0; i < SAMPLES; i++) {
         this->prev_error[i] = 0.0;
-        this->prev_time[i] = ((float) micros()) / ONE_MILLION;
+        this->prev_time[i] = ((double) micros()) / ONE_MILLION;
     }
 }
 
-float PID::compute(float curr_pos) {
-    float curr_time = ((float) micros()) / ONE_MILLION;                                              // get the current time, in seconds
-    float interval_secs = curr_time - this->prev_time[SAMPLES - 1];                                  // compute time passed between loops, in seconds
-    float desired_pos = this->prev_desired_pos + (velocity_to_tps(this->velocity) * interval_secs);  // compute the desired position at this time
-    float error = desired_pos - curr_pos;                                                            // compute the error as the set point (desired position) - process variable (current position)
+double PID::compute(double curr_pos) {
+    double curr_time = ((double) micros()) / ONE_MILLION;                                              // get the current time, in seconds
+    double interval_secs = curr_time - this->prev_time[SAMPLES - 1];                                  // compute time passed between loops, in seconds
+    double desired_pos = this->prev_desired_pos + (velocity_to_tps(this->velocity) * interval_secs);  // compute the desired position at this time
+    double error = desired_pos - curr_pos;                                                            // compute the error as the set point (desired position) - process variable (current position)
     this->integral += error * interval_secs;                                                         // compute the new value of this->integral using right-rectangle approximation
 
     // shift the prev_error and prev_time arrays and insert curr_time and error
@@ -35,7 +35,7 @@ float PID::compute(float curr_pos) {
     this->prev_time[SAMPLES - 1] = curr_time;
 
     // output = kp * error + ki * integral of error * kd * "derivative" of error
-    float output = (this->kp * error) + (this->ki * this->integral) + (this->kd * regression(this->prev_time, this->prev_error, SAMPLES));
+    double output = (this->kp * error) + (this->ki * this->integral) + (this->kd * regression(this->prev_time, this->prev_error, SAMPLES));
 
     // store the current values into previous values for use on next loop
     this->prev_pos = curr_pos;
@@ -51,54 +51,54 @@ float PID::compute(float curr_pos) {
     }
 }
 
-void PID::set_coefficients(float kp, float ki, float kd) {
+void PID::set_coefficients(double kp, double ki, double kd) {
     this->kp = kp;
     this->ki = ki;
     this->kd = kd;
 }
 
-void PID::set_velocity(float velocity) {
+void PID::set_velocity(double velocity) {
     this->velocity = velocity;
 }
 
 // used when student sets encoder to a certain value
 // resets all of the PID controller values
-void PID::set_position(float curr_pos) {
+void PID::set_position(double curr_pos) {
     this->prev_pos = this->prev_desired_pos = curr_pos;
     for (int i = 0; i < SAMPLES - 1; i++) {
         this->prev_error[i] = 0.0;
-        this->prev_time[i] = ((float) micros()) / ONE_MILLION;
+        this->prev_time[i] = ((double) micros()) / ONE_MILLION;
         this->integral = 0.0;
     }
 }
 
 // three getter functions for the three coefficients
-float PID::get_kp() { return this->kp; }
-float PID::get_ki() { return this->ki; }
-float PID::get_kd() { return this->kd; }
+double PID::get_kp() { return this->kp; }
+double PID::get_ki() { return this->ki; }
+double PID::get_kd() { return this->kd; }
 
 // *********************** HELPER FUNCTIONS *********************** //
 
-// find the sum of the given floats
-float PID::sum(float nums[], int i) {
-    float total = 0.0;
+// find the sum of the given doubles
+double PID::sum(double nums[], int i) {
+    double total = 0.0;
     for (int x = 0; x < i; x++) {
         total += nums[x];
     }
     return total;
 }
 
-// find the average of the given floats
-float PID::average(float nums[], int i) {
-    float total = sum(nums, i);
-    return (total / (float) i);
+// find the average of the given doubles
+double PID::average(double nums[], int i) {
+    double total = sum(nums, i);
+    return (total / (double) i);
 }
 
 // calculate the slope of least-squares linear regression through points
 // "number" is the number of elements in the arrays
-float PID::regression(float x_vals[], float y_vals[], int number) {
-    float x_avg, y_avg, slope;
-    float numerator[number], denominator[number];
+double PID::regression(double x_vals[], double y_vals[], int number) {
+    double x_avg, y_avg, slope;
+    double numerator[number], denominator[number];
 
     // calculate averages
     x_avg = average(x_vals, number);
@@ -113,6 +113,6 @@ float PID::regression(float x_vals[], float y_vals[], int number) {
     return slope;
 }
 
-float PID::velocity_to_tps(float velocity) {
+double PID::velocity_to_tps(double velocity) {
     return MAX_TPS * velocity;
 }
